@@ -1,35 +1,19 @@
-// swift-tools-version:5.2
-// The swift-tools-version declares the minimum version of Swift required to build this package.
+// swift-tools-version:5.7
 
 import PackageDescription
 
 let package = Package(
-    name: "GeoJSONKit",
-    products: [
-        // Products define the executables and libraries produced by a package, and make them visible to other packages.
-        .library(
-            name: "GeoJSONKit",
-            targets: ["GeoJSONKit"]),
-    ],
-    dependencies: [
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
-    ],
-    targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
-        .target(
-            name: "GeoJSONKit",
-            dependencies: []),
-        .testTarget(
-            name: "GeoJSONKitTests",
-            dependencies: ["GeoJSONKit"]),
-    ]
+  name: "GeoJSONKit",
+  products: [
+    .library(
+      name: "GeoJSONKit",
+      targets: ["GeoJSONKit"]),
+  ],
+  targets: [
+    .target(
+      name: "GeoJSONKit"),
+    .testTarget(
+      name: "GeoJSONKitTests",
+      dependencies: ["GeoJSONKit"]),
+  ]
 )
-
-#if swift(>=5.6)
-// Add the documentation compiler plugin if possible
-package.dependencies.append(
-  .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0")
-)
-#endif
